+++ src/functions/public/Find-LocalAdsiServerSid.ps1
@@ -1,12 +1,8 @@
 function Find-LocalAdsiServerSid {
 
+    [OutputType([System.String])]
+
     param (
-
-        # Name of the computer to query via CIM
-        [string]$ComputerName,
-
-        # Cache of CIM sessions and instances to reduce connections and queries
-        [hashtable]$CimCache = ([hashtable]::Synchronized(@{})),
 
         <#
         Hostname of the computer running this function.
@@ -14,6 +10,9 @@ function Find-LocalAdsiServerSid {
         Can be provided as a string to avoid calls to HOSTNAME.EXE
         #>
         [string]$ThisHostName = (HOSTNAME.EXE),
+
+        # Name of the computer to query via CIM
+        [string]$ComputerName = $ThisHostName,
 
         <#
         FQDN of the computer running this function.
@@ -25,28 +24,22 @@ function Find-LocalAdsiServerSid {
         # Username to record in log messages (can be passed to Write-LogMsg as a parameter to avoid calling an external process)
         [string]$WhoAmI = (whoami.EXE),
 
-        # Log messages which have not yet been written to disk
-        [Parameter(Mandatory)]
-        [ref]$LogBuffer,
-
         # Output stream to send the log messages to
         [ValidateSet('Silent', 'Quiet', 'Success', 'Debug', 'Verbose', 'Output', 'Host', 'Warning', 'Error', 'Information', $null)]
-        [string]$DebugOutputStream = 'Debug'
+        [string]$DebugOutputStream = 'Debug',
+
+        # In-process cache to reduce calls to other processes or to disk
+        [Parameter(Mandatory)]
+        [ref]$Cache
 
     )
 
-    $Log = @{
-        ThisHostname = $ThisHostname
-        Type         = $DebugOutputStream
-        Buffer       = $LogBuffer
-        WhoAmI       = $WhoAmI
-    }
+    $Log = @{ ThisHostname = $ThisHostname ; Type = $DebugOutputStream ; Buffer = $Cache.Value['LogBuffer'] ; WhoAmI = $WhoAmI }
 
     $CimParams = @{
-        CimCache          = $CimCache
-        ComputerName      = $ThisHostName
+        Cache             = $Cache
+        ComputerName      = $ComputerName
         DebugOutputStream = $DebugOutputStream
-        LogBuffer         = $LogBuffer
         ThisFqdn          = $ThisFqdn
         ThisHostname      = $ThisHostname
         WhoAmI            = $WhoAmI
@@ -61,6 +54,6 @@ function Find-LocalAdsiServerSid {
         return
     }
 
-    return $LocalAdminAccount.SID.Substring(0, $LocalAdminAccount.SID.LastIndexOf("-"))
+    return $LocalAdminAccount.SID.Substring(0, $LocalAdminAccount.SID.LastIndexOf('-'))
 
 }